+++ NEWS.md
@@ -6,6 +6,7 @@ Bugs:
 
 * No warning emitted when `vcov=FALSE` in `lme4` and `glmmTMB` models. Thanks to @ASKurz for report #1460.
 * `inferences()` passes extra arguments to `boot()`, as documented.
+* `inferences(method="rsample")` should not collapse estimates when `term` is not unique (ex: contrast, by, etc.)
 
 ## 0.27.0